+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix `AssetTransfer` module component to not link to the block explorer for native transfers
+
 ### Changed
 
 - Rename `tokenPrice` property to `amountUsd` on `<TransactionDataListItem />` module component